Relax -Wno-error=sign-conversion (clang); cast a few touched sites
Clang's -Wsign-conversion (via -Wconversion, stricter than GCC's) flags
~36 pre-existing benign signed->unsigned conversions (int loop/index vars
used as std::array/span subscripts or size_t args). These were latent on
main behind the earlier -Wmismatched-tags failure. Per project precedent
(the GCC -Wno-error=maybe-uninitialized/format-truncation relaxations),
demote to a non-fatal warning; also keep the handful of explicit casts
already applied while diagnosing. Verified: clang + gcc both build clean
and pass testes/all.lua.

+++ src/compiler/lopcodes.h
@@ -552,7 +552,7 @@ using OpModesArray = std::array<lu_byte, NUM_OPCODES>;
 LUAI_DDEC(const OpModesArray luaP_opmodes;)
 
 inline OpMode getOpMode(int m) noexcept {
-	return static_cast<OpMode>(luaP_opmodes[m] & 7);
+	return static_cast<OpMode>(luaP_opmodes[static_cast<size_t>(m)] & 7);
 }
 
 inline bool testAMode(int m) noexcept {
